Updating utils.collapse_naive_seqs.
Updating to turn off collapsing of sequences with the same inferred
naive sequence, which is done of for Bcell lineages with somatic hyper
mutation. This change should allow partis to be used to calculate the
forward probability (generation probability) of individual Tcell
receptor sequences.

Following advice from Duncan from the Google group:
https://groups.google.com/g/partis/c/BdDP82OjWrQ/m/MNu9o0FmAgAJ